// src/app/Models/resourceaccomplishment.ts
export class resourceaccomplishment {
  id: string;
  buId: any;
  buName: string;
  resourceName: string;
  employeeId: string;
  achievements: string;
  status: string;
  rating: string;
  ratingByname: string;
  ratingbyId: string;
  createdBy: string;
  createdByname: string;
  createdDateTime: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  resultStatus: any;
  message: string;
  roleId: string;
  roleName: string;
  comments: string;
  year: string;
  designation: string;
  reviewedOn:string;
  refId:string;
}